refactor: separate namespace-utils tests
